# transaction pool optimizations