This PR contains some of the commits from the (now reverted) https://github.com/ethereum/go-ethereum/pull/21232 , leaving out the erroneous conversions to `uint64`

```
[user@work core]$ benchstat pool.before pool.after 
name                         old time/op    new time/op    delta
PoolBatchInsert100-6           8.52ms ± 2%    8.50ms ± 2%     ~     (p=1.000 n=5+5)
PoolBatchInsert1000-6          99.2ms ± 1%   101.3ms ± 2%     ~     (p=0.190 n=4+5)
PoolBatchInsert10000-6          1.17s ± 9%     1.17s ±12%     ~     (p=1.000 n=5+5)
PoolBatchLocalInsert100-6       350ms ± 5%      31ms ± 3%  -91.20%  (p=0.008 n=5+5)
PoolBatchLocalInsert1000-6      5.15s ± 8%     0.40s ± 2%  -92.19%  (p=0.016 n=5+4)
PoolBatchLocalInsert10000-6     36.5s ± 2%      3.0s ± 4%  -91.84%  (p=0.036 n=3+5)

name                         old alloc/op   new alloc/op   delta
PoolBatchInsert100-6            293kB ± 0%     264kB ± 0%   -9.71%  (p=0.008 n=5+5)
PoolBatchInsert1000-6          2.63MB ± 0%    2.60MB ± 0%   -1.31%  (p=0.016 n=5+4)
PoolBatchInsert10000-6         22.7MB ± 0%    22.7MB ± 0%   -0.13%  (p=0.008 n=5+5)
PoolBatchLocalInsert100-6      14.7MB ± 0%     0.3MB ± 0%  -97.80%  (p=0.008 n=5+5)
PoolBatchLocalInsert1000-6      205MB ± 8%       3MB ± 0%  -98.71%  (p=0.016 n=5+4)
PoolBatchLocalInsert10000-6    1.46GB ± 0%    0.03GB ± 0%  -98.17%  (p=0.036 n=3+5)

name                         old allocs/op  new allocs/op  delta
PoolBatchInsert100-6            3.78k ± 0%     3.78k ± 0%     ~     (p=0.190 n=5+5)
PoolBatchInsert1000-6           37.8k ± 0%     37.8k ± 0%   -0.16%  (p=0.000 n=5+4)
PoolBatchInsert10000-6           339k ± 0%      339k ± 0%     ~     (p=0.548 n=5+5)
PoolBatchLocalInsert100-6       4.54k ± 0%     3.90k ± 0%  -14.15%  (p=0.016 n=5+4)
PoolBatchLocalInsert1000-6      46.2k ± 1%     38.2k ± 0%  -17.19%  (p=0.008 n=5+5)
PoolBatchLocalInsert10000-6      445k ± 0%      381k ± 0%  -14.36%  (p=0.036 n=3+5)
```